+++ CHANGELOG.md
@@ -131,6 +131,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Fix issue in emulator where holding shift key would cause some inputs not to register
 - Fix issue in emulator where holding inputs while window lost focus could cause buttons to remain stuck
 - Fix issue where text sound effects were not being included in game if they were not referenced by any other events [@pau-tomas](https://github.com/pau-tomas)
+- Fix issue where game may not compile if both Color and Super GB modes are enabled
+- Fix issue where 'Actor Move To' would not always detect one way tile collisions
 
 ## [4.1.3] - 2024-09-16